Send correctly formatted POST-AS-GET in account update
When creating a new account, and account update is requested.

acmev2_req_init is prepared to do the required POST-AS-GET to
retrieve the new account, but md_acme_req_body_init supplies
an empty JSON object instead of the required NULL pointer.

The culprite is on_init_acct_upd(), which needs the JSON
object for ACME V1, but mishandled the V2 case.